move data to new file and reference in user.js
